backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.30
alembic==1.13.1
pydantic==1.10.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.0
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
